Move summariseMatching to tests/helpers folder
