+++ packages/core/tests/helpers/summariseMatching.ts
@@ -1,9 +1,9 @@
 import ExceptionCode from "@moj-bichard7-developers/bichard7-next-data/dist/types/ExceptionCode"
 
-import type { AnnotatedHearingOutcome } from "../../../types/AnnotatedHearingOutcome"
-import type { CourtResultMatchingSummary } from "../../types/MatchingComparisonOutput"
+import type { AnnotatedHearingOutcome } from "../../types/AnnotatedHearingOutcome"
+import type { CourtResultMatchingSummary } from "../types/MatchingComparisonOutput"
 
-import getOffenceCode from "../../../lib/offences/getOffenceCode"
+import getOffenceCode from "../../lib/offences/getOffenceCode"
 
 export const matchingExceptions: ExceptionCode[] = [
   ExceptionCode.HO100203,
