+++ packages/core/tsconfig.build.json
@@ -13,7 +13,7 @@
     "tests/helpers/comparison/generateMockPncQueryResultFromAho.ts",
     "tests/helpers/comparison/getPncQueryTimeFromAho.ts",
     "tests/helpers/parseIncomingMessage.ts",
-    "tests/helpers/comparison/summariseMatching.ts",
+    "tests/helpers/summariseMatching.ts",
     "tests/helpers/comparison/normalisePncOperations.ts",
     "tests/helpers/comparison/processTestFile.ts",
     "tests/types/ComparisonFile.ts",
